PDF report on the EEG exploratory data analysis and it;s implications
This is an EEG DATA analysis. Electroencephalography (EEG) is an electrophysiological neuromonitoring method that measures the weak electromagnetic signals generated by neural activity in the brain. EEG is used to understand different cognitive processes and functions of the brain such as language processing and sensory perception. While more common neuromonitoring methods such as functional magnetic resonance imaging (fMRI) monitors the blood flow to certain areas of the brain thus giving an insight to activities of the neurons in the brain, EEG has a high temporal resolution which captures the changing dynamics of the brain activations at a millisecond time resolution. This is helpful when examining neuronal activity across a wide range of frequencies that can be used for studying contemporary insights into how the brain works as a large system